chore: the logging for development stage is removed from base settings and placed at the .dev settings.
A checklist.txt is also added at algocode_backend/settings/checklist.txt
for a reference of configuration for dev and prod stage, this is a reference if anything related to config is breaking, check the checklist and make chaanges as per the checklist.

Although the development branch holds the dev settings, but it is added for future reference that where does the dev and prod environment differ.